requirements.txt
python-dotenv
selenium
openai
streamlit
requests